feat: Dependency Injection was applied to dao & service instance
