+++ src/main/java/next/controller/HomeController.java
@@ -3,7 +3,7 @@ package next.controller;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-import next.dao.QuestionDao;
+import next.dao.JdbcQuestionDao;
 import core.annotation.Controller;
 import core.annotation.RequestMapping;
 import core.mvc.ModelAndView;
@@ -11,10 +11,10 @@ import core.nmvc.AbstractNewController;
 
 @Controller
 public class HomeController extends AbstractNewController {
-    private QuestionDao questionDao = QuestionDao.getInstance();
+    private JdbcQuestionDao jdbcQuestionDao = JdbcQuestionDao.getInstance();
 
     @RequestMapping("/")
     public ModelAndView execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
-        return jspView("home.jsp").addObject("questions", questionDao.findAll());
+        return jspView("home.jsp").addObject("questions", jdbcQuestionDao.findAll());
     }
 }
