+++ src/main/java/next/controller/qna/ApiListQuestionController.java
@@ -8,7 +8,11 @@ import core.mvc.ModelAndView;
 import next.dao.QuestionDao;
 
 public class ApiListQuestionController extends AbstractController {
-    private QuestionDao questionDao = QuestionDao.getInstance();
+    private QuestionDao questionDao;
+
+    public ApiListQuestionController(QuestionDao jdbcQuestionDao) {
+        this.questionDao = jdbcQuestionDao;
+    }
 
     @Override
     public ModelAndView execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
