+++ src/main/java/next/controller/qna/DeleteAnswerController.java
@@ -10,7 +10,11 @@ import core.mvc.AbstractController;
 import core.mvc.ModelAndView;
 
 public class DeleteAnswerController extends AbstractController {
-    private AnswerDao answerDao = AnswerDao.getInstance();
+    private AnswerDao answerDao;
+
+    public DeleteAnswerController(AnswerDao answerDao) {
+        this.answerDao = answerDao;
+    }
 
     @Override
     public ModelAndView execute(HttpServletRequest request, HttpServletResponse response) throws Exception {
