+++ src/main/java/next/controller/qna/DeleteQuestionController.java
@@ -10,7 +10,11 @@ import next.controller.UserSessionUtils;
 import next.service.QnaService;
 
 public class DeleteQuestionController extends AbstractController {
-    private QnaService qnaService = QnaService.getInstance();
+    private QnaService qnaService;
+
+    public DeleteQuestionController(QnaService qnaService) {
+        this.qnaService = qnaService;
+    }
 
     @Override
     public ModelAndView execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
