+++ src/main/java/next/controller/qna/UpdateQuestionController.java
@@ -10,7 +10,11 @@ import core.mvc.AbstractController;
 import core.mvc.ModelAndView;
 
 public class UpdateQuestionController extends AbstractController {
-    private QuestionDao questionDao = QuestionDao.getInstance();
+    private QuestionDao questionDao;
+
+    public UpdateQuestionController(QuestionDao questionDao) {
+        this.questionDao = questionDao;
+    }
 
     @Override
     public ModelAndView execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
